refactor: organize components with feature folders
