+++ src/components/auth/register/EmailAndPassword.tsx
