+++ src/components/common/AppVersion.tsx
@@ -1,4 +1,4 @@
-import pkg from "../../package.json" assert { type: "json" };
+import pkg from "../../../package.json" assert { type: "json" };
 
 export function AppVersion() {
   return <span>{pkg.version}</span>;
