+++ src/components/icons/AppleIcon.tsx
